Adapting to elastic

+++ tests/testFaucet.js
@@ -11,11 +11,11 @@ require('dotenv').config({
 })
 import config from '../src/config'
 
+const faucetDb = new FaucetDb(config)
+
 chai.use(chaiHttp)
 
 const app = require('../src/server').default
-const faucetDb = new FaucetDb(config)
-
 const { expect } = chai
 let restore
 
@@ -54,9 +54,17 @@ describe('Test: POST /faucet', () => {
         })
     })
 
-    beforeEach(() => {
-        console.log(`Deleting documents from index`)
-        faucetDb.deleteIndexDocuments(config.database.index)
+    beforeEach(async function () {
+        const exists = await faucetDb.indexExists(config.database.index)
+        if (exists) {
+            console.log(`Deleting documents from index`)
+            await faucetDb.deleteIndexDocuments(config.database.index)
+            console.log(`Deleted documents!`)
+        } else {
+            console.log(`Index doesn't exist, creating it`)
+            await faucetDb.initializeIndex(config.database.index)
+        }
+        await faucetDb.refresh(config.database.index)
     })
 
     it('should not POST without an address', (done) => {
@@ -127,9 +135,7 @@ describe('Test: POST /faucet', () => {
             .end((err, res) => {
                 expect(err).to.equal(null)
                 expect(res.body).not.equal(null)
-                expect(res.body.success).to.eql(true)
                 expect(res).to.have.status(201)
-
                 chai.request(app)
                     .post('/faucet')
                     .send(req)
@@ -143,6 +149,31 @@ describe('Test: POST /faucet', () => {
                         expect(res).to.have.status(500)
                         done()
                     })
+            })
+    })
+
+    it('should be able to find a transaction for an existing request', (done) => {
+        const req = {
+            address: '0x1F08a98e53b2bDd0E6aE8E1140017e26E935780D'
+        }
+        let trxHash = ''
+        chai.request(app)
+            .post('/faucet')
+            .send(req)
+            .end((err, res) => {
+                expect(err).to.equal(null)
+                expect(res.body).not.equal(null)
+                expect(res.body.success).to.eql(true)
+                expect(res).to.have.status(201)
+                trxHash = res.body.trxHash
+                done()
+            })
+
+        chai.request(app)
+            .get(`/trxhash?id=${trxHash}`)
+            .end((err, res) => {
+                res.should.have.status(200)
+                done()
             })
     })
